Data in Query can be null

+++ src/test/java/net/nitrado/api/test/GameserverTest.java
@@ -185,14 +185,14 @@ public class GameserverTest extends ApiTestCase {
         Query.Player[] players = query.getPlayers();
         assertEquals(1, players.length);
         Query.Player player = players[0];
-        assertEquals(1, player.getId());
+        assertEquals(1, (int)player.getId());
         assertEquals("Tyrola", player.getName());
         assertFalse(player.isBot());
-        assertEquals(0, player.getScore());
-        assertEquals(0, player.getFrags());
-        assertEquals(0, player.getDeaths());
-        assertEquals(31, player.getTime());
-        assertEquals(8, player.getPing());
+        assertEquals(0, (int)player.getScore());
+        assertEquals(0, (int)player.getFrags());
+        assertEquals(0, (int)player.getDeaths());
+        assertEquals(31,(int)player.getTime());
+        assertEquals(8, (int)player.getPing());
     }
